+++ ts/features/fastLogin/store/selectors/__tests__/fastLogin.test.ts
@@ -2,9 +2,9 @@ import * as O from "fp-ts/lib/Option";
 import { Tuple2, ITuple2 } from "@pagopa/ts-commons/lib/tuples";
 import { BackendStatus } from "../../../../../../definitions/content/BackendStatus";
 import { baseRawBackendStatus } from "../../../../../store/reducers/__mock__/backendStatus";
-import { isNativeLoginEnabledSelector } from "../../../../../store/reducers/backendStatus";
 import { GlobalState } from "../../../../../store/reducers/types";
 import { getAppVersion } from "../../../../../utils/appVersion";
+import { isFastLoginEnabledSelector } from "..";
 
 jest.mock("react-native-device-info", () => ({
   getReadableVersion: jest.fn().mockReturnValue("1.2.3.4"),
@@ -12,18 +12,18 @@ jest.mock("react-native-device-info", () => ({
 }));
 
 jest.mock("../../../../../config", () => ({
-  nativeLoginEnabled: true
+  fastLoginEnabled: true
 }));
 
 const currentAppVersion = getAppVersion();
 
-describe("NativeLogin remote flag test", () => {
+describe("FastLogin remote flag test", () => {
   const status: BackendStatus = {
     ...baseRawBackendStatus
   };
 
-  function checkNativeLoginFlagWithBrokenStatus(expectedValue: boolean) {
-    const customStoreWithMissingMinAppVersionInNativeLoginConfig = {
+  function checkFastLoginFlagWithBrokenStatus(expectedValue: boolean) {
+    const customStoreWithMissingMinAppVersionInFastLoginConfig = {
       backendStatus: {
         status: O.some({
           ...status,
@@ -32,32 +32,31 @@ describe("NativeLogin remote flag test", () => {
       }
     } as unknown as GlobalState;
     const actualStatus =
-      customStoreWithMissingMinAppVersionInNativeLoginConfig.backendStatus
-        .status;
+      customStoreWithMissingMinAppVersionInFastLoginConfig.backendStatus.status;
     expect(O.isSome(actualStatus)).toBe(true);
     if (O.isSome(actualStatus)) {
-      expect(actualStatus.value.config.nativeLogin).toBeUndefined();
+      expect(actualStatus.value.config.fastLogin).toBeUndefined();
     }
-    const isNativeLoginEnabled = isNativeLoginEnabledSelector(
-      customStoreWithMissingMinAppVersionInNativeLoginConfig
+    const isFastLoginEnabled = isFastLoginEnabledSelector(
+      customStoreWithMissingMinAppVersionInFastLoginConfig
     );
-    expect(isNativeLoginEnabled).toBe(expectedValue);
+    expect(isFastLoginEnabled).toBe(expectedValue);
   }
 
-  function checkBrokenNativeLoginFlagTest(
+  function checkBrokenFastLoginFlagTest(
     minAppVersion: string | undefined,
     currentAppVersion: string,
     expectedValue: boolean
   ) {
-    const testTitle = `NativeLogin${
+    const testTitle = `FastLogin${
       expectedValue ? "" : " NOT"
     } enabled with min version ${minAppVersion} for actual version ${currentAppVersion}`;
     it(testTitle, () => {
-      checkNativeLoginFlagWithBrokenStatus(expectedValue);
+      checkFastLoginFlagWithBrokenStatus(expectedValue);
     });
   }
 
-  describe("NativeLogin flag test for empty values", () => {
+  describe("FastLogin flag test for empty values", () => {
     [
       Tuple2("0", false),
       Tuple2("0.0", false),
@@ -77,11 +76,11 @@ describe("NativeLogin remote flag test", () => {
       Tuple2(undefined, false),
       Tuple2("?$&&/!@", false)
     ].forEach((t: ITuple2<string | undefined, boolean>) =>
-      checkBrokenNativeLoginFlagTest(t.e1, currentAppVersion, t.e2)
+      checkBrokenFastLoginFlagTest(t.e1, currentAppVersion, t.e2)
     );
   });
 
-  function checkIfNativeLoginFlagIsEnableForThisAppVersion(
+  function checkIfFastLoginFlagIsEnableForThisAppVersion(
     minAppVersion: string | undefined,
     expectedValue: boolean
   ) {
@@ -91,7 +90,7 @@ describe("NativeLogin remote flag test", () => {
           ...status,
           config: {
             ...status.config,
-            nativeLogin: {
+            fastLogin: {
               min_app_version: {
                 android: minAppVersion,
                 ios: minAppVersion
@@ -102,27 +101,27 @@ describe("NativeLogin remote flag test", () => {
       }
     } as unknown as GlobalState;
 
-    const isNativeLoginEnabled = isNativeLoginEnabledSelector(customStore);
-    expect(isNativeLoginEnabled).toBe(expectedValue);
+    const isFastLoginEnabled = isFastLoginEnabledSelector(customStore);
+    expect(isFastLoginEnabled).toBe(expectedValue);
   }
 
-  function checkNativeLoginFlagTest(
+  function checkFastLoginFlagTest(
     minAppVersion: string | undefined,
     currentAppVersion: string,
     expectedValue: boolean
   ) {
-    const testTitle = `NativeLogin${
+    const testTitle = `FastLogin${
       expectedValue ? "" : " NOT"
     } enabled with min version ${minAppVersion} for actual version ${currentAppVersion}`;
     it(testTitle, () => {
-      checkIfNativeLoginFlagIsEnableForThisAppVersion(
+      checkIfFastLoginFlagIsEnableForThisAppVersion(
         minAppVersion,
         expectedValue
       );
     });
   }
 
-  describe("NativeLogin flag test for different config values", () => {
+  describe("FastLogin flag test for different config values", () => {
     [
       Tuple2("0", false),
       Tuple2("0.0", false),
@@ -142,7 +141,7 @@ describe("NativeLogin remote flag test", () => {
       Tuple2(undefined, false),
       Tuple2("?$&&/!@", false)
     ].forEach((t: ITuple2<string | undefined, boolean>) =>
-      checkNativeLoginFlagTest(t.e1, currentAppVersion, t.e2)
+      checkFastLoginFlagTest(t.e1, currentAppVersion, t.e2)
     );
   });
 });
